fix: update dockerfile for security fixes
It looks like the rhoar-nodejs image **may** not be the most up to date
option for nodejs-10 these days. I built this with the image I"ve added
here and it seems to work. Also added a yum update for packages that
routinely seem to cause security vulns to fire.

Signed-off-by: Stephen Adams <tsadams@gmail.com>