[RHINENG-10134] Remove esrun package and esbuild (its indirect dependency)
In order to fix a Go vulnerability issue caused by `esbuild` which is
deliberately intended to be compatible with an old Go version[1].

This patch removes `esrun` from development dependencies list. `esbuild`
is coming from `esrun` as one of its own dependency, but it looks like
we don't use it at all. So it is safe to remove these packages from
`xjoin-search`.

[1] - https://esbuild.github.io/faq/#old-go-version

Signed-off-by: Gaël Chamoulaud <gchamoul@redhat.com>